+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Web app to test the card layout (built from the actual code with emsdk)
 - Automatic DST adjustment for EEST rules
 - Tested programming and configuration in Windows, added instructions
+- Added arguments to sync-clock.py to change serial port and to execute command script
 
 ### Changed